refactor : VO 객체 PK(id) 타입 변경 int -> Integer

+++ src/main/java/com/sanaiclub/user/model/vo/ClientProfileVO.java
@@ -31,8 +31,8 @@ import lombok.*;
 @Builder
 public class ClientProfileVO {
 
-    private int clientId;
-    private Long companyId;
+    private Integer clientId;
+    private Integer companyId;
     private ClientType clientType;
